remove the uci commit stage from dynapoint
Whenever dynapoint triggers a change to bring up and down certain
wireless interfaces, it is sufficient to just set the uci variable
"disabled" to the appropriate value and do NOT commit this changes,
just reload uci. By removing the actual UCI COMMIT, the changes are
only valid until the next reboot, as they are not written fo /etc/config.
Hence we circumvent different wireless states after rebooting the system.

Signed-off-by: Thomas Huehn <thomas@net.t-labs.tu-berlin.de>